requirements.txt
bitarray>=2.1.0
biopython>=1.78